extract handle jump out logic
